pyproject.toml
[tool.poetry]
name = "stock analysis chatbot"
version = "0.1.0"
description = ""
authors = ["Saran-Kaanth <sarnknth16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
groq = "^0.18.0"
phidata = "^2.7.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
